created NewsSection - Context - Fetch in Main
The fetch renders the entire database, it is necessary to sort the items and limit the quantities, but everything works fine.
